Risk attribution drilldown
Hello!

I've come up with a drilldown to replace the default 'View the individual Risk Attributions' drilldown in a risk incident rule.

The motivation for this was to keep it concise while still providing the unique output that an individual risk rule might provide. 
This can allow an analyst to get the most out of the drilldown before having to pivot.

Open to any and all feedback - especially on _raw data extraction component. I would prefer not to have used regex.. but I couldn't see another way to keep it dynamic.

elusive-mesmer