Some fixes for the stats commands
Removed some unused distinct counts (dc) from the first stats command, as they are not used. Also fixed a small error for the source_count in the second stats. One can not aggregate distinct counts and get a new distinct count, we need the full list of unique sources and do a new distinct count on that list. Also changed so that all risk object "aliases" associated with the "risk object "id" are listed, as opposed to only the latest one. This will also allow the analyst to get the "Risk Event Timeline" for each risk object alias linked to the incident in the Incident Review dashboard.